Move padding into tokenization function, and simplify tokenization
Tokens are now filtered for not being empty outside of the tokenizer struct,
padding is now inside the tokenization function and there's now a big monster type
that defines the results of tokenization. Let's pray for impl Trait.

This does make things slightly nicer and terser though, which is good for everybody.