Significantly speed up training by using a custom joining function
Itertools' one is slow as molasses because it's doing some super crazy
shit in order to be generic. We don't need none of that, and we also know the exact length
of the resulting string as well, so the simpler implementation is much faster.

This also removed a dependency on itertools, which is a good thing.